Expose cert and proxies as attributes of the last request.
When adding timeout and allow_redirects we added the ability to fetch
some of the connection parameters from the request. Given we have made
that link we should also expose the verify, cert and proxies so users
can test that.

Particularly testing that cert has been passed through to a request is
something i've come across a number of times.

Change-Id: I2c60be44769607c32eaf68ad697656b85cae0737