Fix exception message when creating a response
When you register a handler the type for text or data can be a string or
a callback that provides a string. By the time you get to
create_response this callback should have been called and only a string
can be passed.

This check originally would have been in place for when a callback was
run and the return value was not the expected type, however
create_response is a public function that can be used on its own.

Correct the exception to clarify in create_response you must pass a
string or bytes and not a callback.

Change-Id: I3e700afddc1f40454f69a564066495bfc77d91c8
Closes-Bug: #1627506