Reset requests before calling callbacks
When a callback is invoked that calls requests the previous called mock
is run. When this is supposed to be a real_http request the
NoMockAddress is thrown but it can't be handled in the adapter and so
goes back to the outer send function.

This is all a problem and for a really unusual case because almost
surely if you are doing a requests call within a requests-mock callback
you're trying to do a real requests call. There's no point having a
callback that would simply point to something else you registered in the
same adapter.

So let's just assume that and reset all mocking whilst the callback code
may be run.

Change-Id: I9a426c54af778fcd536925744e78843854320a07