Move tests outside of package to top level directory
Cleanly separates tests from the package itself. Prevents the tests
being installed to site-packages. Tests are concerned with development,
not usage.

The tests directory was added to MANIFEST.in, so tests are still
distributed with the source distribution.

This follows the "tests outside application code" convention that is
common among many Python projects and the community.