Another attempt at circleci tags
This was still trying to push to pypi. My guess is that circleci tags
aren't compared with an AND. Having both a branch and tag filter meant
it was attempting to push for anything on master branch.

Change it so that we ignore all branches for publish and only trigger on
a tag.