Support older pytest versions
To be included in EPEL we need to support the current shipped py.test
which is 2.7. This doesn't have the 'bool' type option to addini. We
can copy this handling out of py.test until we can rely on that newer
version.

Also fixes using the @yield_fixture decorator on pytest < 3. This
wouldn't break pytest for other people, but it wouldn't have let you use
it as intended.

BZ: #1605138
Closes: #62