Fix discovery of pytest version and default
So this was a dumb refactoring mistake that was hidden because I was
testing on older versions not newer versions after the refactor.

It means that we were always identifying pytest as an old version, and
(seperately) the default value was then incorrect.

I will scrap the 1.5.1 release and tag this as 1.5.2

Closes: #66