Remove unnecessary compatibility shims
The type bytes is available on all supported Pythons. On Python 2.7, it
is an alias for str, same as six.binary_type.

Likewise, io.BytesIO is available on all supported Pythons. It is always
a stream implementation using an in-memory bytes buffer.

Byte literals and Unicode literals are available on all supported
Pythons.

Makes the code more forward compatible with Python 3.