Mark IOReader object closed when using a stream
When using response.iter_content(None) the path through request is
somewhat different and instead of checking whether the return is empty
it actually checks the closed state of the object.

BytesIO doesn't appear to mark itself as closed automatically when
finished reading so we have to do that ourselves. This doesn't seem to
have any negative effects, but it's deep logic so just have to see.

Fixes: #124