Changed to test to get around linting req
Seems as though the linter requires that methods either use 'self' or
get changed to functions.  In general, I understand the reason.
However, for tests, this is a bit silly.  Nonetheless, the technical
requirement is set up as a gate.  To get around this and because this
project uses multiple asserts per test, I just added an additional
assert.  It is not really useful, but it is harmless.