Remove requests 2.3 compatibility code
The code that parses the requests version is broken when used in
conjunction with PEP440 full versions. It doesn't handle +local and a
number of other cases.

However this code is only being used to check if we're using requests <
2.3 and our requirements have been updated so that we depend on a
requests newer than this.

Therefore instead of fixing the version code lets just remove the
compatability layer.