Revert "Url Quote the path provided to the mocker"
This change broke existing behavior for users who were already quoting
mocked endpoints.

This change also differs slightly from requests's behavior, so it is no
longer possible to register some URIs. For example, if "@" is in the URL
path, requests does not quote this character, but the requests-mock
matcher quotes it to "%40", so the matcher never triggers.

This reverts commit f072845c0cb13c6c0fb18824160639a8bb3c7fe8.