Fix failing pytest
When you start a method with test_ pytest loads it and tries to parse
the parameters as fixtures. In this case it's not a real test, just a
matcher function used later on in the test file.

We can inline it into the function as it's only used in one place.

Fixes: #175